feat: Add image to ban member notification
This commit adds the image of the group to the ban member notification message. Now, when a member is banned from a group, the notification will include the group's image if available. This enhancement improves the visual representation of the notification and provides more context to the user.

Refactor the `banMember` function in `lib/db/group/groupMember.ts` to include the `image` property from the `group` object in the notification message.